auth_backend/backend/requirements.txt
Flask==2.3.2
flask_jwt_extended==4.4.4
pandas==2.0.3
joblib==1.3.1
python-dotenv==1.0.0
